[김서영,김민정] memberId에 따른 자소서,학력사항,자격증 뷰 및 기능 오류 수정

+++ linkJobs/src/main/java/com/gdu/linkJobs/controller/MemberAcademicController.java
@@ -21,9 +21,13 @@ public class MemberAcademicController {
 	
 	//학력 리스트 출력 폼
 	@GetMapping("/getMemberAcademicList")
-	public String getMemberAcademic(HttpSession session,Model model,
-			@RequestParam("memberId") String memberId) {
-		
+	public String getMemberAcademic(HttpSession session,Model model) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
+		String memberId=(String)session.getAttribute("loginMember");
+		MemberAcademic MA=new MemberAcademic();
+		MA.setMemberId(memberId);
 		List<MemberAcademic> academicList = academicService.getMemberAcademicList(memberId);
 		model.addAttribute("academicList", academicList);
 		
@@ -33,20 +37,30 @@ public class MemberAcademicController {
 	//학력 추가 폼
 	@GetMapping("/addMemberAcademic")
 	public String addMemberAcademic(HttpSession session) {
-		
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		return "member/addMemberAcademic";
 	}
 	//학력 추가 액션
 	@PostMapping("/addMemberAcademic")
 	public String addMemberAcademic(HttpSession session,MemberAcademic memberAcademic) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
+		String memberId=(String)session.getAttribute("loginMember");
+		memberAcademic.setMemberId(memberId);
 		academicService.addMemberAcademic(memberAcademic);
-		return "redirect:/member/addMemberAcademic";
+		return "redirect:/getMemberAcademicList";
 	}
 	
 	//학력 수정 폼
 	@GetMapping("/modifyMemberAcademic")
 	public String modifyMemberAcademic(HttpSession session,Model model,
 			@RequestParam("academicNo") int academicNo) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		MemberAcademic ma=academicService.getMemberAcademicOne(academicNo);
 		model.addAttribute("ma", ma);
 		return "member/modifyMemberAcademic";
@@ -55,15 +69,23 @@ public class MemberAcademicController {
 	//학력 수정 액션
 	@PostMapping("/modifyMemberAcademic")
 	public String modifyMemberAcademic(HttpSession session,MemberAcademic memberAcademic) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
+		
 		academicService.modifyMemberAcademic(memberAcademic);
-		return "redirect:/member/modifyMemberAcademic";
+		return "redirect:/getMemberAcademicList";
 	}
 	
 	//학력 삭제 액션
 	@GetMapping("/removeMemberAcademic")
-	public String removeMemberAcademic(HttpSession session,@RequestParam("academicNo") int academicNo) {
+	public String removeMemberAcademic(HttpSession session, @RequestParam("academicNo") int academicNo) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
+		System.out.println(academicNo+"<--academicNo.removeMemberAcademic");
 		academicService.removeMemberAcademic(academicNo);
-		return "redirect:/member/removeMemberAcademic";
+		return "redirect:/getMemberAcademicList";
 	}
